refactor: 后端健壮性修复 + 测试覆盖补全
- CORS 移除 allow_credentials=True（与通配 origin 矛盾）
- get_device_id 用 Header(None) 惯用法
- 模型 avg_price 用 Decimal 诚实标注；schema stores 用 default_factory
- repository 抽 _keyword_filter、补 order_by、新增 get_by_ids
- 收藏幂等加 IntegrityError 兜底、列表 N+1 改批量查询
- AI 抽 DeepSeekError 精准映射 502（DB 错误不再误报）；兜底改逐类独立
- 种子用 2.0 select、幂等同时检查 Food/Attraction
- 测试 23→36：分页边界/空结果/描述搜索/收藏景点分支/status=True/AI 兜底与空引用

+++ backend/app/modules/food/model.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, ForeignKey, Numeric, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -13,7 +14,7 @@ class Food(Base):
     id: Mapped[int] = mapped_column(BigIntType, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(String(100), nullable=False)
     description: Mapped[str] = mapped_column(Text, nullable=False)
-    avg_price: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    avg_price: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     image_url: Mapped[str | None] = mapped_column(String(255), nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
     updated_at: Mapped[datetime] = mapped_column(
